Add import with public address and ens domain (#3384)
* Add import with public address

This changes the import flow so that the user can just share a public address, instead of
giving out secret phrase. When the wallet is in this state, the user can see funds and NFTs,
they can connect to dApps, but they can't sign in or transact. If they try to make any of these
we show a toast prompting them to import credentials. There is a visible "Import Wallet credentials"
button inside the wallet panel. Also we lock UnlockedThemes section with a similar prompt.

Otherwise, all functionality remains the same. If the user imports credentials, we update the related
bits and they can use all functionality.

* Add ens domain support and a ways to import using ens domains

* Failure for import and version bump

* set assets # limit to 50

* Use the right ens name

* Add pagination to asset fetch

* Version bump again and fix assetview buttons again

* Fix toast for importing wallet credentials