Fix send eth (#3571)
Fix the send flow by moving to a NavigationLink from actionSheet. Looks like actionSheet
is not interacting with TextField well, so since the send form will need to include text entry
it is more meaningful to use a NavigationLink here. I had to move things around a bit to make this
work and also open the other menu as NavigationLink as well