Avoid sync call back to main thread to write out preference (#4181)
Just make the call directly from the background thread. Use UserDefaults
directly instead of going through the @Default property wrapper. Remove
the other use of @Default as well since this is not a SwiftUI View, and
we can avoid the overhead of hooking up an observer for that preference.

Fixes #4173