Add caching to FaviconResolver to avoid flickering of favicons (#4202)
The system frequently creates and recreates FaviconView objects,
and the FaviconResolver has an async DB lookup to find the URL.
As a result, each time a FaviconView gets created it would show
the fallback content momentarily before showing the right image.
This can result in a fair bit of flicker.

The simple solution here is to maintain an in-memory cache of
site URLs to favicon URLs. This way we can synchronously know
the favicon URL to use and avoid showing fallback content in
most cases.

It's necessary to update this cache whenever FaviconHandler
learns of a new mapping.

This also seems to fix the issue with the favicon of a card
animating out of sync with the rest of the card during the
card transition animation. This is probably some side-effect
of animating the view while it is transitioning from one view
to another. Maybe tickling a bug in SwiftUI or something, but
the cache helps avoid this issue too. (Except on iOS 16, where
there is still an issue of this sort to work out.)

Fixes #3632, #3551