Fix crash by writing to UserDefaults on the main thread (#4244)
Should be no need to use BackgroundTaskProcessor when we are already on
the main thread.

The SO article mentioned in https://github.com/neevaco/neeva-ios/pull/3667
was about trying to write to UserDefaults from a background thread. We have
many other cases where we write to UserDefaults on the main thread during
shutdown and have not observed data loss.

Fixes #4243