Move archived tabs out of TabManager.tabs (#4253)
With this change, operations on tabs is now independent of the
number of archived tabs, which should improve performance.

Introduces ArchivedTab type and GenericTab as a protocol shared by
Tab and ArchivedTab. ArchivedTab is a thin wrapper around SavedTab.

Doesn't change the way we persist tabs, including archived tabs,
but does pave the way for that as a follow-up.

Also flattens out the directory structure under Browser/Tabs to be more
manageable on the command line :-)

A side-effect of this PR is that the createOrSwitchToTab method will no longer
consider archived tabs. That's a performance win and paves the way for these
not needing to be kept in memory, but it is also a change in behavior. The plan
is to experiment with this behavior change and see how it works out.