Reduce various expensive updates that happen while showing the card grid (#4347)
Avoid calling TabCardModel.updateRows() when taking a new screenshot of the tab, which happens as we go to show the card grid. We avoid this by having TabCardDetails monitor the published Tab.screenshotUUID instead. Because Card observes the CardDetails, this results in the Card picking up the new screenshot.

Avoid calling updateAllTabDataAndSendNotifications when just selecting the tab that is already selected. Showing the card grid and then showing the same tab again should not need to re-generate all of the tabs data. This is done by adding a needsHeavyUpdatesPostAnimation flag to TabManager that gets set when selectTab would otherwise do this heavy work.

Avoid watching all of GridModel in various layers of the code. A number of properties of GridModel get poked during the transition to the card grid, and by breaking these out into their own model classes, we can avoid updating so much UI when making changes to a subset of those properties. This translated to creating GridResizeModel, GridScrollModel and GridSwitcherModel.

Switch to just observing some particular published vars on some models (e.g., GridScrollModel) instead of the entire model to avoid spurious UI updates as well. Add a useEffect variant to make this easier.

Move observation of TabChromeModel out of CardScrollContainer and into EmptyCardGrid where it is needed, which helps cut down on invalidations of CardScrollContainer.

The net result of this is that for a card grid with a single card, we reduce the number of FaviconView updates by about 30%.

Issue #4342