Fix TabGridContainer.selectRowId implementation (#4387)
When deciding to select the heading instead of the row, be sure to select the right heading (not just the pinned section). This requires searching for the heading within the set of rows.

Also be sure to actually set the Row.id on each generate row HStack. This way we can scroll to the specified row.

This fixes a bug where users would sometimes not be able to see the card grid when tapping on the "show tabs" button.

Also fixes the case of a param of getLogCounterAttributesForTabGroups.

Fixes #4386