Eliminate one update of CardsContainer when transitioning to / from card grid (#4382)
Move CardsContainer's dependency on GridVisibilityModifier into a modifier to isolate the impact of updates to that model. This dependency is just used to ensure that the spaces view is showing the first card. This change cuts down the number of card grid updates from 3 to 2 when transitioning to / from the card grid.

Other small improvements in this PR include ensuring that we only update cardSize when it changes. That way we avoid more unnecessary updates to SwiftUI. This PR also eliminates an unused containerGeom state variable.

Also fixes it so that scrolling to the top-most space works when the enableSpaceDigestCard experiment is disabled by using SpaceCardViewModel.dataSource instead of SpaceCardModel.allDetails.

Issue #4364