Stop using WebImage for FaviconView to improve performance on iOS 16 (#4395)
There was some change in the iOS 16 SwiftUI implementation that has led to
issues rendering favicons in the card grid. They always start out by showing
fallback content before loading the actual image.

To workaround this issue we can just implement our own image fetcher and
cache. The cache enables us to synchronously provide a UIImage to SwiftUI,
which eliminates any flicker.

This change also helps avoid the issue of the favicon moving out of sync with
the card during a card grid transition. It doesn't seem to completely eliminate
that issue however.

This change should also help make the card grid transition smoother overall.
However, I have not measured the impact.

The cache is written to in two places: 1) when loading a web page and fetching
the favicon the first time, we store the image in this new cache, and 2) when
there is a cache miss and we need to fetch the image again.

Issue #4364