Eliminate unnecessary spring animation added to scrollview fling (#4459)
When scrolling the page up or down and hitting the extents of the
scrollview, you will often see the scrollview jitter a bit as it
hits those bounds. It does not smoothly rubberband like normal
scrollviews. This is due to the `withAnimation` around the assignment
of `headerTopOffset` and `footerBottomOffset`.

Fixes #2696