Add ability to create Space from Tab Group (#4494)
This feature allows a user to create a Space from a Tab Group. The Tab Group's name is pulled directly from the name of the Tab Group. Also, the user is NOT able to add a Tab Group to an existing Space (this will be implemented later).

The UI for this feature is behind a feature flag.