Correctly handle user-submitted custom scheme queries (#4526)
Previously, when a user submitted a custom scheme URL (like lyft://book-a-ride) into the address bar, Neeva would submit a search. In order to ensure feature parity with Safari iOS, this was changed so that the corresponding app will be launched (if it is installed on the device). If the corresponding app is not on the device, Neeva will fall back to submitting a search.