Add Tab Group to existing Space (#4541)
Previously, with the feature flag on, users could create a Tab Group from a Space. However, they could not add a Tab Group to an existing Space. This commit adds this functionality. It can be accessed through the overflow menu of any Tab Group. The feature flag has also been removed, meaning this will now be available to all users.

On the more technical side, a struct called AddToSpaceInput was created to simplify the transfer of data during these operations.